Update GXDLMSProfileGeneric.cs
The datetime.MinValue causes a problem when a GMT + offset is used. An exception is thrown when comparing the lastDate to the MinValue. Since the code only tries to determine if there is a valid value in the lastDate variable, I changed it to use a Nullable. That way, no comparison is needed with the MinValue and no exception is thrown.
This is the exception btw: The UTC time represented when the offset is applied must be between year 0 and 10,000. Parameter name: offset